requirements.txt
accelerate~=0.25.0
datasets~=2.15.0
loguru~=0.7.2
numpy~=1.26.2
openai~=1.6.0
pandas~=2.1.4
pymilvus~=2.3.4
python-dotenv~=1.0.0
rich~=13.7.0
scipy~=1.11.4
scikit-learn~=1.3.2
tiktoken~=0.5.2
torch~=2.1.2
tqdm~=4.66.1
transformers~=4.36.2
typer~=0.9.0